% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI510FinalProj-package.R
\docType{package}
\name{BMI510FinalProj}
\alias{BMI510FinalProj-package}
\alias{BMI510FinalProj}
\title{BMI510FinalProj}
\description{
`BMI510-FinalProj` is an R package that completes the requirements for the BMI 510 course's final project, on behalf JaMor Hairston.
}
\details{
\itemize{
  \item \code{\link{logLikBernoulli}}: Calculates the parameter that maximizes the log-likelihood for Bernoulli-distributed data.
  \item \code{\link{survCurv}}: Generates and plots survival curves.
  \item \code{\link{unscale}}: Reverses centering and scaling transformations on data.
  \item \code{\link{pcApprox}}: Reconstructs a dataset using principal components.
  \item \code{\link{standardizeNames}}: Standardizes variable names in a dataset to smallCamelCase.
  \item \code{\link{minimumN}}: Calculates the minimum sample size required for a t-test with specified power and alpha.
  \item \code{\link{downloadRedcapReport}}: Downloads a report from REDCap using an API.
}
}
\author{
JaMor Hairston \email{jamor.hairston@emory.edu}
}
