% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReort.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download REDCap Report}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable storing the REDCap API token.}

\item{redcapUrl}{The base URL of the REDCap API.}

\item{redcapReportId}{The identifier for the specific REDCap report to download.}
}
\value{
A tibble containing the downloaded REDCap report.
}
\description{
This function downloads a report from REDCap using API parameters and returns it as a tibble.
The given API token is read from an environment variable for security.
}
\examples{
# Ensure that REDCap API token is stored in .Renviron as "MY_REDCAP_TOKEN"
downloadRedcapReport("MY_REDCAP_TOKEN", "https://redcap.example.com/api/", "12345")
}
