% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculate the maximizing parameter p for Bernoulli log-likelihood}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{Numeric vector containing Bernoulli trial outcomes (0s and 1s).}
}
\value{
A list containing the maximum log-likelihood and the corresponding value of p.
}
\description{
This function performs a grid search to find the parameter p that maximizes the
log-likelihood of a Bernoulli-distributed dataset. It assumes p varies between 0 and 1,
and increments by 0.001 for each step in the search.
}
\examples{
data = c(1, 0, 0, 0, 1, 1, 1)
logLikBernoulli(data)
}
