% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Calculate Min Sample Size for a t-Test}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{A numeric vector of preliminary data for the one-sample or the first sample in a two-sample t-test.}

\item{x2}{An optional numeric vector of preliminary data for the second sample in a two-sample t-test.}
}
\value{
The minimum sample size required to achieve the specified power and significance level.
}
\description{
This function calculates the minimum sample size needed for a one-sample or two-sample t-test
on the provided data to achieve a power of 80% and alpha of 0.05.
}
\examples{
# One-sample t-test
data1 = rnorm(10, mean = 5)
minimumN(data1)

# Two-sample t-test
data2 = rnorm(10, mean = 5.5)
minimumN(data1, data2)
}
