% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Column Names to smallCamelCase}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble whose column names are to be standardized.}
}
\value{
A tibble with standardized column names.
}
\description{
This function takes a tibble and standardizes its variable names to smallCamelCase
using `dplyr::rename_with`, `janitor::make_clean_names`, and `snakecase::to_small_camel_case`.
}
\examples{
library(tibble)
data = tibble(`Column one` = 1:5, `COLUMN TWO` = letters[1:5])
standardized_data = standardizeNames(data)
names(standardized_data)
}
