% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Plot Survival Curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{Numeric vector indicating the status (1 = event occurred, 0 = censored).}

\item{time}{Numeric vector indicating the time to event or censorship.}
}
\value{
The plot of the survival curve.
}
\description{
This function takes numerical vectors for status and time, and plots the survival curve S(t).
}
\examples{
survCurv(status, time)
}
