% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Reverse Scaling of a Vector}
\usage{
unscale(x, original_mean, original_sd)
}
\arguments{
\item{x}{A numeric vector that has been scaled.}

\item{original_mean}{The mean used to center the data before scaling.}

\item{original_sd}{The standard deviation used to scale the data.}
}
\value{
A numeric vector with the scaling reversed.
}
\description{
This function takes a scaled vector and reverses the scaling using the original mean and standard deviation.
}
\examples{
original_vector = c(1, 2, 3, 4, 5)
scaled_vector = scale(original_vector)
unscaled_vector = unscale(scaled_vector, mean(original_vector), sd(original_vector))
all.equal(original_vector, unscaled_vector)  # Should return TRUE
}
